allow radio to access IOemSlsiRadioExternal hwservice
It's needed for system_ext/app/OemRilHookService which runs in the radio domain.

OemRilHookService publishes "telephony.oem.oemrilhook" service and tries to connect to
IOemSlsiRadioExternal when handling calls to this service.

In some cases, OemRilHookService started to retry access to IOemSlsiRadioExternal tens of times per
minute for multiple hours, which led to excessive battery drain.